moved userName state up to App level
